Fix: [CI] Resolve Doxygen warnings, MSVC conversions, unused strings, and JSON newlines.
- Doxygen: Add single-line @file header and comprehensive member/parameter documentation to src/portal/balancing_critic.h.
- MSVC: Use size_t in Planetary Stockpile loop in src/portal/balancing_critic.cpp to prevent C4267 data conversion warnings.
- Reports: Add missing trailing newlines to reports/bom_tuning_patch.json and reports/critic_20yr.json.
- Strings: Clean unused STR_ERROR_FABRICATION_RESEARCH_OR_CASH from english.txt, and wire up StringID constants in corporate_hq_gui.cpp and universe_directory_gui.cpp.

+++ src/portal/balancing_critic.cpp
@@ -318,15 +318,15 @@ BalancingCriticReport BalancingCritic::AnalyzeAndCritique(const std::vector<Bala
 	}
 
 	/* 4. Planetary Stockpile Starvation */
-	uint32_t total_stock_checks = 0;
-	uint32_t zero_stock_checks = 0;
+	size_t total_stock_checks = 0;
+	size_t zero_stock_checks = 0;
 	for (const auto &snap : timeseries) {
 		for (const auto &s : snap.stockpiles) {
 			total_stock_checks += s.inventory.size();
 			zero_stock_checks  += s.zero_stock_cargos;
 		}
 	}
-	report.stockpile_starvation_rate = (total_stock_checks > 0) ? (static_cast<double>(zero_stock_checks) / total_stock_checks) : 0.0;
+	report.stockpile_starvation_rate = (total_stock_checks > 0) ? (static_cast<double>(zero_stock_checks) / static_cast<double>(total_stock_checks)) : 0.0;
 
 	if (report.stockpile_starvation_rate > 0.25) {
 		report.diagnoses.push_back(fmt::format(
